+++ plugins/outputs/influxdb/influxdb.go
@@ -48,7 +48,7 @@ var sampleConfig = `
   urls = ["http://localhost:8086"] # required
   ### The target database for metrics (telegraf will create it if not exists)
   database = "telegraf" # required
-  ### Precision of writes, valid values are n, u, ms, s, m, and h
+  ### Precision of writes, valid values are "ns", "us" (or "µs"), "ms", "s", "m", "h".
   ### note: using "s" precision greatly improves InfluxDB compression
   precision = "s"
 
@@ -156,17 +156,20 @@ func (i *InfluxDB) Description() string {
 // Choose a random server in the cluster to write to until a successful write
 // occurs, logging each unsuccessful. If all servers fail, return error.
 func (i *InfluxDB) Write(metrics []telegraf.Metric) error {
-	bp, _ := client.NewBatchPoints(client.BatchPointsConfig{
+	bp, err := client.NewBatchPoints(client.BatchPointsConfig{
 		Database:  i.Database,
 		Precision: i.Precision,
 	})
+	if err != nil {
+		return err
+	}
 
 	for _, metric := range metrics {
 		bp.AddPoint(metric.Point())
 	}
 
 	// This will get set to nil if a successful write occurs
-	err := errors.New("Could not write to any InfluxDB server in cluster")
+	err = errors.New("Could not write to any InfluxDB server in cluster")
 
 	p := rand.Perm(len(i.conns))
 	for _, n := range p {